Re-entrance for bbl up.
- The ops-file flag takes a path to an ops-file. We will write the
current state.BOSH.UserOpsFile to a file as the default for that flag.
If the user would like to remove the current ops-file applied, they will
need to pass an empty ops-file via the command-line flag.
- The credhub flag must be opted into on the first bbl up, otherwise it
will return an error similar to how we do not let you change the name of
your environment.
- Added missing IAAS up config test coverage.

[#150500100]

Signed-off-by: Genevieve LEsperance <glesperance@pivotal.io>