Revert cf public http health check name change
- In #145893715, we needed a new HealthCheck on the router lb backend
service because the protocol and port name were now HTTPS. The existing
HttpHealthCheck was repurposed for just the ws target pool. The change
from HttpHealthCheck -> HealthCheck was forcing terraform to recreate
the health checks which led to resourceInUseByAnotherResource errors.
- We reverted the name of the existing health check and added a new
health check for the router lb backend service in order to mitigate
any errors for users migrating from 3.X to 4.2.2

[#150646260]

Signed-off-by: Genevieve LEsperance <glesperance@pivotal.io>