Call bosh create-env with args from interpolate
Rather than calling bosh create-env and bosh delete-env
with an already-interpolated manifest, call it with the
manifest and all the args that we use when we call interpolate
(vars-file, ops-files, etc).

We no longer check the bosh-state.json file directly or write
it back into the bbl-state.json file.  Instead, we just rely
on the state file being written to the location specified in
these respective calls to bosh.

[#151614247]

Signed-off-by: Rowan Jacobs <rojacobs@pivotal.io>