Write GCP service account key to BOSH vars-files
Ideally, we would not do this, but we are currently writing
IAAS credentials for *all* IAASes to the BOSH vars files.

In order to not do this, we need to pass them along with the
`--var` flag like we are doing for terraform now, but this will
require some larger changes. For instance, we may need to rethink
how we are writing out the args to create-env as shell scripts.

[#153755946]

Signed-off-by: Dave McClure <dmcclure@pivotal.io>