Fix bosh-lite-gcp plan patch
Renames file as `bosh-lite.tf`, as all the resources are new and
therefore this cannot be an override file in Terraform v1.4.

Removes the quotes from the depends_on assignment, as that's deprecated
in Terraform v1.4.

Removes the `external_ip` output as it's a duplicate of the
`external_ip` output in `bbl-template.tf`.